move lambda API utils to localstack.testing
